testdatapublishing: Test Data Publishing yes yes yes Test Data Publishing yes yes yes This is testing publishing the data Test Data Publishing yes yes yes

testdatapublishing: Test Data Publishing yes yes yes Test Data Publishing yes yes yes This is testing publishing the data Test Data Publishing yes yes yes

Usage

testdatapublishing

Format

A tibble with 11448 rows and 18 variables

country

NA

date_submitted

NA

photo_of_the_water_point

NA

latitude

NA

longtude

NA

altitude

NA

functional_status

NA

current_problem

NA

other_current_problems

NA

water_quality_issues_with_this_water_point

NA

water_quality_issues_on_the_water_point

NA

other_water_quality_issues_on_the_water_point

NA

households_who_use_this_water_point

NA

are_there_times_of_the_year_when_water_is_not_available_from_this_source_due_to_seasonal_variation

NA

when_is_water_not_available_from_this_source

NA

is_there_a_service_provider_or_someone_responsible_for_operating_and/or_maintaining_this_water_point_or_water_system

NA

type_of_service_provider

NA

water_point_have_active_water_point_committee

NA